Add ST_3DLength
Note 1: PostGIS does not allow ST_3DLength to take a Polygon as a
parameter. For this purpose, a PostGIS user must use ST_Perimeter, which
was not implemented in GDMS.

Note 2: PostGIS also allows MutliLineStrings (for ST_3DLength) and
MultiPolygons (for ST_Perimeter), while GDMS only allowed for
LineStrings and Polygons.